+++ cypress/elements/dashboardItem.js
@@ -1,5 +1,3 @@
-import { EXTENDED_TIMEOUT } from '../support/utils.js'
-
 export const chartClass = '.highcharts-container'
 export const chartSubtitleClass = '.highcharts-subtitle'
 export const chartXAxisLabelClass = '.highcharts-xaxis-labels'
@@ -13,18 +11,3 @@ export const itemMenuButtonSel = 'dashboarditem-menu-button'
 export const mapLegendButtonClass = '.dhis2-map-legend-button'
 export const mapLegendContentClass = '.dhis2-map-legend-content'
 export const mapLegendTitleClass = '.dhis2-map-legend-title'
-
-export const getDashboardItem = (itemUid) =>
-    cy.getBySel(`dashboarditem-${itemUid}`, EXTENDED_TIMEOUT)
-
-export const clickItemMenuButton = (itemUid) =>
-    getDashboardItem(itemUid)
-        .scrollIntoView()
-        .findBySel(itemMenuButtonSel)
-        .click()
-
-export const clickItemDeleteButton = (itemUid) =>
-    getDashboardItem(itemUid)
-        .scrollIntoView()
-        .findBySel('delete-item-button')
-        .click()
