--- cypress/elements/sharingDialog.js
@@ -1,2 +1,0 @@
-export const getSharingDialogUserSearch = () =>
-    cy.get('[placeholder="Enter names"]').scrollIntoView()
